+++ n-api/lib/bindings.js
@@ -4,5 +4,5 @@ var path = require('path');
 var binding_path = binary.find(path.resolve(path.join(__dirname,'../package.json')));
 var binding = require(binding_path);
 
-//module.exports = require('../build/Debug/crossfilter.node');
-module.exports = binding;
+module.exports = require('../build/Debug/crossfilter.node');
+//module.exports = binding;
